start extracting target-only merge result
